updating pjas program to take average and newcsv to not overwrite

+++ newcsv.py
@@ -1,5 +1,6 @@
 import datetime
 import csv
+import os
 
 class NewCsv:
     
@@ -12,9 +13,15 @@ class NewCsv:
         newfilename = f'wether_data_{self.filetime}.csv'
         
         if self.filetime not in filename:
-            with open(newfilename, 'w') as csvfile:
-                csvfileWriter = csv.DictWriter(csvfile, ['Time', 'Number of drops per frame', 'rain on ground'])
-                csvfileWriter.writeheader()
-            return newfilename
+            if os.path.exists(newfilename):
+                with open(newfilename, 'a') as csvfile:
+                    csvfileWriter = csv.DictWriter(csvfile, ['Time', 'Number of drops per frame', 'rain on ground'])
+                return newfilename
+        
+            else:
+                with open(newfilename, 'w') as csvfile:
+                    csvfileWriter = csv.DictWriter(csvfile, ['Time', 'Number of drops per frame', 'rain on ground'])
+                    csvfileWriter.writeheader()
+                return newfilename
         else:
             return filename
